add small database
