+++ home_interface/views.py
@@ -1,7 +1,15 @@
 from django.shortcuts import render, HttpResponse, redirect
+from django.shortcuts import render, get_object_or_404
+from .models import User, Competition, Team
 
-
-
+def new_competition():
+    new_competition = Competition(
+        name="Competition 1 from sql",
+        start_date="2023-11-01",
+        end_date="2023-11-30",
+        description="Description of the new competition."
+    )
+    new_competition.save()
 
 # Create your views here.
 def home(request):
@@ -11,4 +19,5 @@ def blog(request):
     return render(request, "blog.html")
 
 def blogs(request):
-    return render(request, "blog-single.html")
+    competition = Competition.objects.first()
+    return render(request, 'blog-single.html', {'competition': competition})
